[MOD] implemented %d output with width
Why: To print width-fixed integers and pass timetest.

How: Using a bias method when dealing with %d syntax. The integer will
be saved in the last part of vbuf, and the vsprintf will return the bias
so that strcat could find the start and work. Length is calculated using
the bias now.

Tags: printf %d stdio klib